Fixes -- group folders (partial), save-in-place
When saving, it shows an error, but that is related to the download
doesn't happen. It saves in place.

Old versions of the file are available in Details --> Versions.

Group folders are only partially fixed in that a user named "testuser"
must be created on the system and granted access to the group files.

This will be fixed by using webdav for file retrieval instead of api.